pyproject.toml
[project]
name = "compose-me"
version = "0.1.0"
description = "Helm's little sister for Docker Compose."
authors = [
    {name = "Niklas Rosenstein", email = "rosensteinniklas@gmail.com"},
]
dependencies = [
    "pyyaml>=6.0.1",
    "jinja2>=3.1.2",
    "typer>=0.9.0",
    "mypy>=1.7.1",
    "black>=23.11.0",
    "flake8>=6.1.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
compose-me = "compose_me.__main__:app"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "mypy>=1.7.0",
    "flake8>=6.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "pytest>=7.4.3",
    "types-pyyaml>=6.0.12.12",
    "mksync>=0.1.4",
    "mkdocs>=1.5.3",
    "mkdocs-nav-weight>=0.2.0",
]

[tool.pdm.scripts]
lint = "bash -xc 'mypy src && flake8 src --max-line-length 120 --ignore=W503,W504,E203 && black src --check && isort src --check-only'"
fmt = "bash -xc 'black src && isort src && (mksync -i README.md || true)'"
test = "pytest src test -vv"
docs = "bash -c 'cd docs && mkdocs serve'"
"docs:build" = "bash -c 'cd docs && mkdocs build'"

# Linter/Formatter configuration
# ------------------------------

[tool.mypy]
explicit_package_bases = true
mypy_path = ["src"]
exclude = ["templates"]
namespace_packages = true
pretty = true
python_version = "3.10"
show_error_codes = true
show_error_context = true
strict = true
warn_no_return = true
warn_redundant_casts = true
warn_unreachable = true
warn_unused_ignores = true

[tool.isort]
combine_as_imports = true
line_length = 120
profile = "black"

[tool.black]
line-length = 120
